+++ demos/manual/src/main/java/is/codion/framework/demos/manual/keybinding/KeyBindingModel.java
@@ -74,8 +74,8 @@ final class KeyBindingModel {
 
 	private void bindEvents(FilterComboBoxModel<Item<LookAndFeelProvider>> lookAndFeelComboBoxModel) {
 		componentComboBoxModel.refresher().success().addListener(tableModel::refresh);
-		componentComboBoxModel.selectionChanged().addListener(tableModel::refresh);
-		lookAndFeelComboBoxModel.selectionChanged().addListener(componentComboBoxModel::refresh);
+		componentComboBoxModel.selectedItem().addListener(tableModel::refresh);
+		lookAndFeelComboBoxModel.selectedItem().addListener(componentComboBoxModel::refresh);
 	}
 
 	private static String className(String componentName) {