+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/component/table/FilterColumnConditionPanel.java
@@ -574,7 +574,7 @@ public final class FilterColumnConditionPanel<C, T> extends ColumnConditionPanel
 						.maximumRowCount(operators.size())
 						.mouseWheelScrollingWithWrapAround(true)
 						.toolTipText(conditionModel().operator().get().description())
-						.onBuild(comboBox -> operatorComboBoxModel.selectionChanged().addConsumer(selectedOperator ->
+						.onBuild(comboBox -> operatorComboBoxModel.selectedItem().addConsumer(selectedOperator ->
 										comboBox.setToolTipText(selectedOperator.value().description())))
 						.build();
 	}